Update general_commands.rs
When a user chose to reveal a song when viewing it's properties, it would open the file with the default music app instead of revealing it in the file manager. The issue has now been addressed on windows and macos, however for linux based systems, since there is no universal command, the issue will remain unfixed.
If a user using a linux based distro anticipates they will be using a specific file manager, they can build a binary with the specific name and command embedded into this file.